+++ client/scripts/views/pages/treasury.ts
@@ -12,7 +12,7 @@ import Substrate from 'controllers/chain/substrate/main';
 
 import Sublayout from 'views/sublayout';
 import { PageLoading } from 'views/pages/loading';
-import { ProposalCard } from 'views/components/proposal_card';
+import { ProposalCard } from 'views/components/proposal_card/proposal_card';
 import { CountdownUntilBlock } from 'views/components/countdown';
 import Listing from 'views/pages/listing';
 import ErrorPage from 'views/pages/error';
