+++ client/scripts/views/pages/view_proposal/header.ts
@@ -25,7 +25,6 @@ import {
 import { ProposalType } from 'types';
 
 import { notifyError, notifySuccess } from 'controllers/app/notifications';
-import { getStatusClass, getStatusText } from 'views/components/proposal_card';
 
 import { confirmationModalWithText } from 'views/modals/confirm_modal';
 import { alertModalWithText } from 'views/modals/alert_modal';
@@ -35,6 +34,10 @@ import { activeQuillEditorHasText, GlobalStatus } from './body';
 import { IProposalPageState } from '.';
 import OffchainVotingModal from '../../modals/offchain_voting_modal';
 import { CWIcon } from '../../components/component_kit/cw_icons/cw_icon';
+import {
+  getStatusClass,
+  getStatusText,
+} from '../../components/proposal_card/helpers';
 
 export const ProposalHeaderExternalLink: m.Component<{
   proposal: AnyProposal | OffchainThread;
